matched .jsx to .tsx extention
