+++ src/app/layout.tsx
@@ -1,47 +1,44 @@
-
-import { Geist, Geist_Mono, Rajdhani } from "next/font/google";
-import "./globals.css";
-import { NavBar } from "../components/navbar"
-import { ThemeProvider } from "../components/theme-provider";
+import { Geist, Geist_Mono, Rajdhani } from 'next/font/google';
+import './globals.css';
+import NavBar from '../components/navbar';
+import { ThemeProvider } from '../components/theme-provider';
 
 //Font Family
 const geistSans = Geist({
-  variable: "--font-geist-sans",
-  subsets: ["latin"],
+  variable: '--font-geist-sans',
+  subsets: ['latin'],
 });
 const geistMono = Geist_Mono({
-  variable: "--font-geist-mono",
-  subsets: ["latin"],
+  variable: '--font-geist-mono',
+  subsets: ['latin'],
 });
 const rajdhani = Rajdhani({
-  subsets: ["latin"],
-  weight: ["400", "700"],
-  variable: "--font-rajdhani",
+  subsets: ['latin'],
+  weight: ['400', '700'],
+  variable: '--font-rajdhani',
 });
 
-
 export const metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
+  title: 'Create Next App',
+  description: 'Generated by create next app',
 };
 
 export default function RootLayout({ children }: { children: React.ReactNode }) {
   return (
     <html lang="en" suppressHydrationWarning>
-      <body 
+      <body
         className={`${geistSans.variable} ${geistMono.variable} ${rajdhani.variable} antialiased`} //Font Family
       >
         <ThemeProvider
-            attribute="class"
-            defaultTheme="light"
-            enableSystem
-            disableTransitionOnChange
+          attribute="class"
+          defaultTheme="light"
+          enableSystem
+          disableTransitionOnChange
         >
-          <NavBar/>
+          <NavBar />
           {children}
         </ThemeProvider>
       </body>
     </html>
   );
 }
-
